backend/requirements.txt
# Compatible versions for LUSID SDK
lusid-sdk==2.0.456
pydantic>=1.10.5,<2.0.0
fastapi>=0.68.0,<0.100.0
uvicorn[standard]>=0.15.0
requests>=2.25.0
python-dotenv>=0.19.0
